+++ src/py/app/domain/accounts/schemas/_roles.py
@@ -37,5 +37,5 @@ class RoleCreate(CamelizedBaseStruct):
     name: str
 
 
-class RoleUpdate(CamelizedBaseStruct):
+class RoleUpdate(CamelizedBaseStruct, omit_defaults=True):
     name: str | msgspec.UnsetType | None = msgspec.UNSET